Began reworking structure
