+++ TODocumentPickerViewController/Controllers/TODocumentPickerTableViewController.h
@@ -21,9 +21,18 @@
 //  IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 #import <UIKit/UIKit.h>
-#import "TODocumentPickerViewController.h"
+#import "TODocumentPickerConstants.h"
 
-@interface TODocumentPickerTableViewController : UITableViewController
+
+@interface TODocumentPickerViewController : UITableViewController
+
+@property (nonatomic, strong) TODocumentPickerViewControllerDataSource *dataSource;              /* Data source for file info. Retained by the document picker. */
+@property (nonatomic, weak)   id<TODocumentPickerViewControllerDelegate> documentPickerDelegate; /* Sends out delegate events to the assigned object */
+
+@property (nonatomic, strong) NSArray *allowedFileExtensions;   /* File formats that may be selected by this controller. (If nil, all files may be selected) */
+@property (nonatomic, assign) BOOL showExcludedFileExtensions;  /* Shows files that weren't on the allowed extensions list, but grayed out (Default is NO) */
+@property (nonatomic, strong) NSDictionary *fileFormatIcons;    /* If dealing with custom formats, this lets you add custom icons for those formats (Images must be 40x40 points) */
+@property (nonatomic, strong) NSDictionary *themeAttributes;    /* Attributes for applying a new colour scheme to this view controller. */
 
 @property (nonatomic, copy)   void (^refreshControlTriggeredHandler)(void); /* Handler block called whenever the refresh control is triggered. */
 @property (nonatomic, copy)   NSString *filePath; /* The file path that this view controller corresponds to */
@@ -33,6 +42,8 @@
 @property (nonatomic, strong) UIImage *folderIcon;
 @property (nonatomic, strong) NSDictionary *fileIcons;
 @property (nonatomic, assign) Class tableViewCellClass;
+
+@property (nonatomic, copy, readonly) void (^updateItemsForFilePath)(NSArray *items, NSString *filePath);
 
 @end
 
@@ -46,3 +57,16 @@
 - (void)pushNewViewControllerForFilePath:(NSString *)filePath animated:(BOOL)animated;
 
 @end
+
+//-------------------------------------------------------------------------
+// Theming Attributes
+
+extern NSString *const TODocumentPickerViewControllerThemeAttributeBackgroundColor;                     /* Background color of the table view */
+extern NSString *const TODocumentPickerViewControllerThemeAttributeTableSeparatorColor;                 /* Color of the table cell divider lines */
+extern NSString *const TODocumentPickerViewControllerThemeAttributeTableCellTitleColor;                 /* Color of the title text label in each cell */
+extern NSString *const TODocumentPickerViewControllerThemeAttributeTableCellDetailTextColor;            /* Color of the subtitle text label in each cell */
+extern NSString *const TODocumentPickerViewControllerThemeAttributeTableCellAccessoryTintColor;         /* Color of the arrow accessory icon */
+extern NSString *const TODocumentPickerViewControllerThemeAttributeTableCellIconTintColor;              /* Tint color of the icons in each cell */
+extern NSString *const TODocumentPickerViewControllerThemeAttributeTableSectionHeaderBackgroundColor;   /* Background color of each section header */
+extern NSString *const TODocumentPickerViewControllerThemeAttributeTableSectionTitleColor;              /* Color of the text in each section header */
+extern NSString *const TODocumentPickerViewControllerThemeAttributeTableSectionIndexColor;              /* Tint color of the scrollable section index column */
