+++ TODocumentPickerViewController/TODocumentPickerConstants.h
@@ -1,5 +1,5 @@
 //
-//  TODocumentPickerViewController.h
+//  TODocumentPickerDefines.h
 //
 //  Copyright 2015 Timothy Oliver. All rights reserved.
 //
@@ -21,41 +21,53 @@
 //  IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 #import <UIKit/UIKit.h>
-#import "TODocumentPickerItem.h"
+#import <Foundation/FOundation.h>
 
-@class TODocumentPickerViewControllerDataSource;
-@protocol TODocumentPickerViewControllerDelegate;
+@class TODocumentPickerViewController;
 
-@interface TODocumentPickerViewController : UINavigationController
+/*
+ The default content styles that the controller implements.
+ */
+typedef NS_ENUM(NSInteger, TODocumentPickerViewControllerStyle) {
+    TODocumentPickerViewControllerStyleLightContent,
+    TODocumentPickerViewControllerStyleDarkContent
+};
 
-@property (nonatomic, strong) TODocumentPickerViewControllerDataSource *dataSource;              /* Data source for file info. Retained by the document picker. */
-@property (nonatomic, weak)   id<TODocumentPickerViewControllerDelegate> documentPickerDelegate; /* Sends out delegate events to the assigned object */
-
-@property (nonatomic, strong) NSArray *allowedFileExtensions;   /* File formats that may be selected by this controller. (If nil, all files may be selected) */
-@property (nonatomic, assign) BOOL showExcludedFileExtensions;  /* Shows files that weren't on the allowed extensions list, but grayed out (Default is NO) */
-@property (nonatomic, strong) NSDictionary *fileFormatIcons;    /* If dealing with custom formats, this lets you add custom icons for those formats (Images must be 40x40 points) */
-@property (nonatomic, strong) NSDictionary *themeAttributes;    /* Attributes for applying a new colour scheme to this view controller. */
-
-@end
+/* 
+ The various sorting orders in which the picker can display the files. 
+*/
+typedef NS_ENUM(NSInteger, TODocumentPickerSortType) {
+    TODocumentPickerSortTypeNameAscending,
+    TODocumentPickerSortTypeNameDescending,
+    TODocumentPickerSortTypeDateAscending,
+    TODocumentPickerSortTypeDateDescending,
+    TODocumentPickerSortTypeSizeAscending,
+    TODocumentPickerSortTypeSizeDescending
+};
 
 //-------------------------------------------------------------------------
-// An abstract class used to provide the document picker with per-folder file information
+// The data source protocol for objects charged with loading the file information to display
 
-@interface TODocumentPickerViewControllerDataSource : NSObject
+@protocol TODocumentPickerViewControllerDataSource <NSObject>
+
+/**
+ The first instance of the document view controller that this data source object was assigned.
+ */
+@property (nonatomic, readonly) TODocumentPickerViewController *rootDocumentPickerViewController;
 
 /**
  The class for which table view cells in the document picker will be instantiated off.
  This can be used to insert new table view cell classes with additional configurable views.
  If this property is not set, the default value of [UITableViewCell Class] is used instead
  */
-@property (nonatomic, assign) Class tableViewCellClass;
+- (Class)tableCellViewClassForDocumentPickerViewController:(TODocumentPickerViewController *)documentPicker;
 
 /**
  Whether synchronously or asynchronously, when your data source has finished building
  a list of items, you must call this block to pass those items back up to the view controller.
  (For UI considerations, this block always executes on the main thread)
  */
-@property (nonatomic, copy, readonly) void (^updateItemsForFilePath)(NSArray *items, NSString *filePath);
+//@property (nonatomic, copy, readonly) void (^updateItemsForFilePath)(NSArray *items, NSString *filePath);
 
 /**
  Called by the view controller when it wants to obtain a list of items for the folder at the end of the file path.
@@ -64,7 +76,7 @@
  
  @param filePath The file path with which to download file information
  */
-- (void)requestItemsForFilePath:(NSString *)filePath;
+- (void)documentPickerViewController:(TODocumentPickerViewController *)documentPicker requestItemsForFilePath:(NSString *)filePath;
 
 /**
  If an asynchronous request for files is currently progress and the representing view controller is canceled,
@@ -72,13 +84,13 @@
  
  @param filePath The file path with which to cancel loading
  */
-- (void)cancelRequestForFilePath:(NSString *)filePath;
+- (void)documentPickerViewController:(TODocumentPickerViewController *)documentPicker cancelRequestForFilePath:(NSString *)filePath;
 
 /**
  The title that will appear in the navigation bar for the folder at this file path.
  If not implemented, defautl behaviour is to return simply the folder name from the filePath string
  */
-- (NSString *)titleForFilePath:(NSString *)filePath;
+- (NSString *)documentPickerViewController:(TODocumentPickerViewController *)documentPicker titleForFilePath:(NSString *)filePath;
 
 @end
 
@@ -92,18 +104,4 @@
 /* User selected a set of items and tapped */
 - (void)documentPickerViewController:(TODocumentPickerViewController *)documentPicker didSelectItems:(NSArray *)items;
 
-@end
-
 //-------------------------------------------------------------------------
-// Theming Attributes
-
-extern NSString *const TODocumentPickerViewControllerThemeAttributeBackgroundColor;                     /* Background color of the table view */
-extern NSString *const TODocumentPickerViewControllerThemeAttributeTableSeparatorColor;                 /* Color of the table cell divider lines */
-extern NSString *const TODocumentPickerViewControllerThemeAttributeTableCellTitleColor;                 /* Color of the title text label in each cell */
-extern NSString *const TODocumentPickerViewControllerThemeAttributeTableCellDetailTextColor;            /* Color of the subtitle text label in each cell */
-extern NSString *const TODocumentPickerViewControllerThemeAttributeTableCellAccessoryTintColor;         /* Color of the arrow accessory icon */
-extern NSString *const TODocumentPickerViewControllerThemeAttributeTableCellIconTintColor;              /* Tint color of the icons in each cell */
-extern NSString *const TODocumentPickerViewControllerThemeAttributeTableSectionHeaderBackgroundColor;   /* Background color of each section header */
-extern NSString *const TODocumentPickerViewControllerThemeAttributeTableSectionTitleColor;              /* Color of the text in each section header */
-extern NSString *const TODocumentPickerViewControllerThemeAttributeTableSectionIndexColor;              /* Tint color of the scrollable section index column */
-
